+++ R/plot-combo-response_curve.R
@@ -101,8 +101,8 @@ plot_dose_response_combo <- function(dt_average,
   dt_avg <- dt_avg[selected_combination, on = c(cellline_name, drug_name, drug_name_2)]
 
   # Ensure the drug with more dose levels is on the x-axis
-  n_conc_1 <- length(unique(dt_avg[[conc]][dt_avg[[conc]] > 0]))
-  n_conc_2 <- length(unique(dt_avg[[conc_2]][dt_avg[[conc_2]] > 0]))
+  n_conc_1 <- sum(unique(dt_avg[[conc]]) > 0)
+  n_conc_2 <- sum(unique(dt_avg[[conc_2]]) > 0)
   if (n_conc_2 > n_conc_1) {
     tmp <- dt_avg[[conc]]
     data.table::set(dt_avg, j = conc, value = dt_avg[[conc_2]])
@@ -256,13 +256,13 @@ plot_dose_response_combo_panel <- function(dt_average,
 
   # Ensure the drug with more dose levels is on the x-axis for each pair
   pairs <- unique(dt_avg[, .SD, .SDcols = c(drug_name, drug_name_2)])
-  for (i in seq_len(NROW(pairs))) {
+  for (i in seq_len(nrow(pairs))) {
     d1 <- pairs[[drug_name]][i]
     d2 <- pairs[[drug_name_2]][i]
     idx <- which(dt_avg[[drug_name]] == d1 & dt_avg[[drug_name_2]] == d2)
-    n_c1 <- length(unique(dt_avg[[conc]][idx][dt_avg[[conc]][idx] > 0]))
-    n_c2 <- length(unique(dt_avg[[conc_2]][idx][dt_avg[[conc_2]][idx] > 0]))
-    if (n_c2 > n_c1) {
+    n_conc_1 <- sum(unique(dt_avg[[conc]][idx]) > 0)
+    n_conc_2 <- sum(unique(dt_avg[[conc_2]][idx]) > 0)
+    if (n_conc_2 > n_conc_1) {
       old_conc <- dt_avg[[conc]][idx]
       old_conc_2 <- dt_avg[[conc_2]][idx]
       old_d <- dt_avg[[drug_name]][idx]
@@ -306,10 +306,10 @@ plot_dose_response_combo_panel <- function(dt_average,
     ggplot2::scale_x_log10(oob = scales::squish_infinite) +
     ggplot2::scale_color_manual(values = colormap,
                                 labels = levels(ls_conc_2)) +
-    ggplot2::xlab(bquote(~ "Concentration of Drug [" ~ mu * M ~ "]")) +
+    ggplot2::xlab(bquote(~ "Concentration [" ~ mu * M ~ "]")) +
     ggplot2::ylab(normalization_type) +
     ggplot2::ggtitle(panel_title) +
-    ggplot2::labs(color = bquote(~ "Codrug [" ~ mu * M ~ "]")) +
+    ggplot2::labs(color = bquote(~ "Co-treatment [" ~ mu * M ~ "]")) +
     ggplot2::theme_bw() +
     ggplot2::theme(
       axis.text.x = ggplot2::element_text(size = 8, angle = 45, vjust = 1, hjust = 1, margin = ggplot2::margin()),